+++ ydb/core/tx/columnshard/counters/portions.cpp
@@ -4,14 +4,14 @@
 namespace NKikimr::NColumnShard {
 
 void TPortionCategoryCounters::AddPortion(const std::shared_ptr<NOlap::TPortionInfo>& p) {
-    RecordsCount->Add(p->NumRows());
+    RecordsCount->Add(p->GetRecordsCount());
     Count->Add(1);
     BlobBytes->Add(p->GetTotalBlobBytes());
     RawBytes->Add(p->GetTotalRawBytes());
 }
 
 void TPortionCategoryCounters::RemovePortion(const std::shared_ptr<NOlap::TPortionInfo>& p) {
-    RecordsCount->Remove(p->NumRows());
+    RecordsCount->Remove(p->GetRecordsCount());
     Count->Remove(1);
     BlobBytes->Remove(p->GetTotalBlobBytes());
     RawBytes->Remove(p->GetTotalRawBytes());
@@ -29,7 +29,7 @@ void TSimplePortionsGroupInfo::AddPortion(const TPortionInfo& p) {
     BlobBytes += p.GetTotalBlobBytes();
     RawBytes += p.GetTotalRawBytes();
     Count += 1;
-    RecordsCount += p.NumRows();
+    RecordsCount += p.GetRecordsCount();
     ChunksCount += p.GetChunksCount();
 }
 
@@ -41,7 +41,7 @@ void TSimplePortionsGroupInfo::RemovePortion(const TPortionInfo& p) {
     BlobBytes -= p.GetTotalBlobBytes();
     RawBytes -= p.GetTotalRawBytes();
     Count -= 1;
-    RecordsCount -= p.NumRows();
+    RecordsCount -= p.GetRecordsCount();
     ChunksCount -= p.GetChunksCount();
     AFL_VERIFY(RawBytes >= 0);
     AFL_VERIFY(BlobBytes >= 0);
