+++ ydb/core/tx/columnshard/engines/portions/portion_info.cpp
@@ -1,5 +1,6 @@
 #include "column_record.h"
 #include "constructor.h"
+#include "data_accessor.h"
 #include "portion_info.h"
 
 #include <ydb/core/tx/columnshard/data_sharing/protos/data.pb.h>
@@ -8,24 +9,6 @@
 #include <ydb/core/tx/columnshard/engines/storage/chunks/data.h>
 
 namespace NKikimr::NOlap {
-
-ui64 TPortionInfo::GetColumnRawBytes(const std::set<ui32>& entityIds, const bool validation) const {
-    ui64 sum = 0;
-    const auto aggr = [&](const TColumnRecord& r) {
-        sum += r.GetMeta().GetRawBytes();
-    };
-    AggregateIndexChunksData(aggr, Records, &entityIds, validation);
-    return sum;
-}
-
-ui64 TPortionInfo::GetColumnBlobBytes(const std::set<ui32>& entityIds, const bool validation) const {
-    ui64 sum = 0;
-    const auto aggr = [&](const TColumnRecord& r) {
-        sum += r.GetBlobRange().GetSize();
-    };
-    AggregateIndexChunksData(aggr, Records, &entityIds, validation);
-    return sum;
-}
 
 ui64 TPortionInfo::GetColumnRawBytes() const {
     AFL_VERIFY(Precalculated);
@@ -37,28 +20,10 @@ ui64 TPortionInfo::GetColumnBlobBytes() const {
     return PrecalculatedColumnBlobBytes;
 }
 
-ui64 TPortionInfo::GetIndexRawBytes(const std::set<ui32>& entityIds, const bool validation) const {
-    ui64 sum = 0;
-    const auto aggr = [&](const TIndexChunk& r) {
-        sum += r.GetRawBytes();
-    };
-    AggregateIndexChunksData(aggr, Indexes, &entityIds, validation);
-    return sum;
-}
-
-ui64 TPortionInfo::GetIndexRawBytes(const bool validation) const {
-    ui64 sum = 0;
-    const auto aggr = [&](const TIndexChunk& r) {
-        sum += r.GetRawBytes();
-    };
-    AggregateIndexChunksData(aggr, Indexes, nullptr, validation);
-    return sum;
-}
-
 TString TPortionInfo::DebugString(const bool withDetails) const {
     TStringBuilder sb;
     sb << "(portion_id:" << PortionId << ";"
-       << "path_id:" << PathId << ";records_count:" << NumRows()
+       << "path_id:" << PathId << ";records_count:" << GetRecordsCount()
        << ";"
           "min_schema_snapshot:("
        << MinSnapshotDeprecated.DebugString()
@@ -92,18 +57,6 @@ TString TPortionInfo::DebugString(const bool withDetails) const {
     return sb << ")";
 }
 
-std::vector<const NKikimr::NOlap::TColumnRecord*> TPortionInfo::GetColumnChunksPointers(const ui32 columnId) const {
-    std::vector<const TColumnRecord*> result;
-    for (auto&& c : Records) {
-        if (c.ColumnId == columnId) {
-            Y_ABORT_UNLESS(c.Chunk == result.size());
-            Y_ABORT_UNLESS(c.GetMeta().GetNumRows());
-            result.emplace_back(&c);
-        }
-    }
-    return result;
-}
-
 void TPortionInfo::RemoveFromDatabase(IDbWrapper& db) const {
     db.ErasePortion(*this);
     for (auto& record : Records) {
@@ -125,75 +78,6 @@ void TPortionInfo::SaveToDatabase(IDbWrapper& db, const ui32 firstPKColumnId, co
             db.WriteIndex(*this, record);
         }
     }
-}
-
-std::vector<NKikimr::NOlap::TPortionInfo::TPage> TPortionInfo::BuildPages() const {
-    std::vector<TPage> pages;
-    struct TPart {
-    public:
-        const TColumnRecord* Record = nullptr;
-        const TIndexChunk* Index = nullptr;
-        const ui32 RecordsCount;
-        TPart(const TColumnRecord* record, const ui32 recordsCount)
-            : Record(record)
-            , RecordsCount(recordsCount) {
-        }
-        TPart(const TIndexChunk* record, const ui32 recordsCount)
-            : Index(record)
-            , RecordsCount(recordsCount) {
-        }
-    };
-    std::map<ui32, std::deque<TPart>> entities;
-    std::map<ui32, ui32> currentCursor;
-    ui32 currentSize = 0;
-    ui32 currentId = 0;
-    for (auto&& i : Records) {
-        if (currentId != i.GetColumnId()) {
-            currentSize = 0;
-            currentId = i.GetColumnId();
-        }
-        currentSize += i.GetMeta().GetNumRows();
-        ++currentCursor[currentSize];
-        entities[i.GetColumnId()].emplace_back(&i, i.GetMeta().GetNumRows());
-    }
-    for (auto&& i : Indexes) {
-        if (currentId != i.GetIndexId()) {
-            currentSize = 0;
-            currentId = i.GetIndexId();
-        }
-        currentSize += i.GetRecordsCount();
-        ++currentCursor[currentSize];
-        entities[i.GetIndexId()].emplace_back(&i, i.GetRecordsCount());
-    }
-    const ui32 entitiesCount = entities.size();
-    ui32 predCount = 0;
-    for (auto&& i : currentCursor) {
-        if (i.second != entitiesCount) {
-            continue;
-        }
-        std::vector<const TColumnRecord*> records;
-        std::vector<const TIndexChunk*> indexes;
-        for (auto&& c : entities) {
-            ui32 readyCount = 0;
-            while (readyCount < i.first - predCount && c.second.size()) {
-                if (c.second.front().Record) {
-                    records.emplace_back(c.second.front().Record);
-                } else {
-                    AFL_VERIFY(c.second.front().Index);
-                    indexes.emplace_back(c.second.front().Index);
-                }
-                readyCount += c.second.front().RecordsCount;
-                c.second.pop_front();
-            }
-            AFL_VERIFY(readyCount == i.first - predCount)("ready", readyCount)("cursor", i.first)("pred_cursor", predCount);
-        }
-        pages.emplace_back(std::move(records), std::move(indexes), i.first - predCount);
-        predCount = i.first;
-    }
-    for (auto&& i : entities) {
-        AFL_VERIFY(i.second.empty());
-    }
-    return pages;
 }
 
 ui64 TPortionInfo::GetMetadataMemorySize() const {
@@ -350,8 +234,8 @@ void TPortionInfo::ReorderChunks() {
 }
 
 void TPortionInfo::FullValidation() const {
-    CheckChunksOrder(Records);
-    CheckChunksOrder(Indexes);
+    TPortionDataAccessor::CheckChunksOrder(Records);
+    TPortionDataAccessor::CheckChunksOrder(Indexes);
     AFL_VERIFY(PathId);
     AFL_VERIFY(PortionId);
     AFL_VERIFY(MinSnapshotDeprecated.Valid());
@@ -370,84 +254,6 @@ void TPortionInfo::FullValidation() const {
     } else {
         AFL_VERIFY(blobIdxs.empty());
     }
-}
-
-ui64 TPortionInfo::GetMinMemoryForReadColumns(const std::optional<std::set<ui32>>& columnIds) const {
-    ui32 columnId = 0;
-    ui32 chunkIdx = 0;
-
-    struct TDelta {
-        i64 BlobBytes = 0;
-        i64 RawBytes = 0;
-        void operator+=(const TDelta& add) {
-            BlobBytes += add.BlobBytes;
-            RawBytes += add.RawBytes;
-        }
-    };
-
-    std::map<ui64, TDelta> diffByPositions;
-    ui64 position = 0;
-    ui64 RawBytesCurrent = 0;
-    ui64 BlobBytesCurrent = 0;
-    std::optional<ui32> recordsCount;
-
-    const auto doFlushColumn = [&]() {
-        if (!recordsCount && position) {
-            recordsCount = position;
-        } else {
-            AFL_VERIFY(*recordsCount == position);
-        }
-        if (position) {
-            TDelta delta;
-            delta.RawBytes = -1 * RawBytesCurrent;
-            delta.BlobBytes = -1 * BlobBytesCurrent;
-            diffByPositions[position] += delta;
-        }
-        position = 0;
-        chunkIdx = 0;
-        RawBytesCurrent = 0;
-        BlobBytesCurrent = 0;
-    };
-
-    for (auto&& i : Records) {
-        if (columnIds && !columnIds->contains(i.GetColumnId())) {
-            continue;
-        }
-        if (columnId != i.GetColumnId()) {
-            if (columnId) {
-                doFlushColumn();
-            }
-            AFL_VERIFY(i.GetColumnId() > columnId);
-            AFL_VERIFY(i.GetChunkIdx() == 0);
-            columnId = i.GetColumnId();
-        } else {
-            AFL_VERIFY(i.GetChunkIdx() == chunkIdx + 1);
-        }
-        chunkIdx = i.GetChunkIdx();
-        TDelta delta;
-        delta.RawBytes = -1 * RawBytesCurrent + i.GetMeta().GetRawBytes();
-        delta.BlobBytes = -1 * BlobBytesCurrent + i.GetBlobRange().Size;
-        diffByPositions[position] += delta;
-        position += i.GetMeta().GetNumRows();
-        RawBytesCurrent = i.GetMeta().GetRawBytes();
-        BlobBytesCurrent = i.GetBlobRange().Size;
-    }
-    if (columnId) {
-        doFlushColumn();
-    }
-    i64 maxRawBytes = 0;
-    TDelta current;
-    for (auto&& i : diffByPositions) {
-        current += i.second;
-        AFL_VERIFY(current.BlobBytes >= 0);
-        AFL_VERIFY(current.RawBytes >= 0);
-        if (maxRawBytes < current.RawBytes) {
-            maxRawBytes = current.RawBytes;
-        }
-    }
-    AFL_VERIFY(current.BlobBytes == 0)("real", current.BlobBytes);
-    AFL_VERIFY(current.RawBytes == 0)("real", current.RawBytes);
-    return maxRawBytes;
 }
 
 ISnapshotSchema::TPtr TPortionInfo::TSchemaCursor::GetSchema(const TPortionInfoConstructor& portion) {
@@ -486,7 +292,7 @@ void TPortionInfo::Precalculate() {
             PrecalculatedColumnRawBytes += r.GetMeta().GetRawBytes();
             PrecalculatedColumnBlobBytes += r.BlobRange.GetSize();
         };
-        AggregateIndexChunksData(aggr, Records, nullptr, true);
+        TPortionDataAccessor::AggregateIndexChunksData(aggr, Records, nullptr, true);
     }
 }
