+++ ydb/core/tx/columnshard/engines/storage/actualizer/counters/counters.h
@@ -34,13 +34,13 @@ public:
     }
 
     void AddPortion(const std::shared_ptr<TPortionInfo>& p) {
-        RecordsCount->Add(p->NumRows());
+        RecordsCount->Add(p->GetRecordsCount());
         Count->Add(1);
         Bytes->Add(p->GetTotalBlobBytes());
     }
 
     void RemovePortion(const std::shared_ptr<TPortionInfo>& p) {
-        RecordsCount->Remove(p->NumRows());
+        RecordsCount->Remove(p->GetRecordsCount());
         Count->Remove(1);
         Bytes->Remove(p->GetTotalBlobBytes());
     }
