+++ ydb/core/tx/columnshard/splitter/abstract/chunk_meta.h
@@ -14,7 +14,7 @@ namespace NKikimr::NOlap {
 
 class TSimpleChunkMeta {
 protected:
-    ui32 NumRows = 0;
+    ui32 RecordsCount = 0;
     ui32 RawBytes = 0;
     TSimpleChunkMeta() = default;
 public:
@@ -24,11 +24,8 @@ public:
         return sizeof(ui32) + sizeof(ui32);
     }
 
-    ui32 GetNumRows() const {
-        return NumRows;
-    }
     ui32 GetRecordsCount() const {
-        return NumRows;
+        return RecordsCount;
     }
     ui32 GetRawBytes() const {
         return RawBytes;
